% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chicago}
\alias{chicago}
\title{Chicago crime data}
\format{
A \code{data.frame} with 720 rows and 4 variables:
\describe{
  \item{year}{4 digit year}
  \item{month}{abbreviation for month of observation}
  \item{type}{type of crime}
  \item{count}{number of reported crimes}
}
}
\source{
\url{}
}
\usage{
chicago
}
\description{
Add details about this data set here
}
\keyword{datasets}
