% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_analysis.R
\name{dlm_analysis}
\alias{dlm_analysis}
\title{Function to perform analysis of crime data using dynamic linear models}
\usage{
dlm_analysis(data, prior, ...)
}
\arguments{
\item{data}{a \code{data.frame} containing aggregated crime data}

\item{prior}{a list with elements ... containing a prior specification,
see details}

\item{...}{passed to \code{sampling} in rstan}
}
\value{
a \code{data.frame} containing posterior samples for model parameters
}
\description{
Add more detailed information here,
see http://r-pkgs.had.co.nz/man.html for details on using roxygen2 for
documentation.
}
\details{
Provide model and prior specification here.
}
\examples{
\dontrun{
dlm_analysis(chicago, list(a=1))
}

}
\keyword{internal}
