% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_means.R
\name{get_means}
\alias{get_means}
\title{Function to compute the estimated time series means.}
\usage{
get_means(states, harmonics = 4, crime_types = c("burglary", "robbery"))
}
\arguments{
\item{states}{A three dimensional array of sampled states such as that given by \code{get_states}.}

\item{harmonics}{An integer in 1 to 6 specifiying the number of harmonics to model seasonality.}

\item{crime_types}{Character vector specifying the types of crimes to include in the analysis.
The function requires at least two crime types.}
}
\value{
Returns a three dimensional array where dimension one corresponds to the sample/iteration,
dimension two corresponds to the number of time points, and dimension three corresponds to the crime type.
}
\description{
\code{get_means} simply computes the linear combination of parameters defined to be the data model mean for a given set of sampled states.
}
\examples{
 \dontrun{
 cov_samples <- run_mcmc(data = chicago, chains = 2, adapt_delta = 0.8)
 state_samples <- get_states(mcmc_samples = cov_samples$samples, data = chicago)
 ts_means <- get_means(states = state_samples)
 }
}
