% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonality.R
\name{get_seasonality}
\alias{get_seasonality}
\title{Function to compute the estimated seasonal time series components.}
\usage{
get_seasonality(states, harmonics = 4, crime_types = c("burglary", "robbery"))
}
\arguments{
\item{states}{A three dimensional array of sampled states such as that given by \code{get_states}.}

\item{harmonics}{An integer in 1 to 6 specifiying the number of harmonics to model seasonality.
Must be consistent with specification in \code{run_mcmc()}.}

\item{crime_types}{Character vector specifying the types of crimes to include in the analysis.
The function requires at least two crime types.}
}
\value{
Returns a list with two objects: seasonal_effects and crime_types.
seasonal_effects is a three dimensional array where dimension one corresponds to the sample/iteration,
dimension two corresponds to the number of time points, and dimension three corresponds to the crime type.
crime_types returns the crime_types argument specified in the function. This serves as a reminder of
the ordering of crime types, which is important for interpreting the smoothmeans output
as well as making sure that the order of crime types is consistent between other functions.
}
\description{
\code{get_seasonality} computes the seasonal component in the observation equation of the program narrative.
}
\examples{
 \dontrun{
 cov_samples <- run_mcmc(data = chicago, chains = 2, adapt_delta = 0.8)
 state_samples <- get_states(mcmc_samples = cov_samples$samples, data = chicago)
 seasonality <- get_seasonality(states = state_samples$state_samples)
 }
}
