% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Function to sample observation and state covariance parameters.}
\usage{
run_mcmc(data, harmonics = 4, chains = 3, iter = 2000, warmup = 1000,
  adapt_delta = 0.9, crime_types = c("burglary", "robbery"),
  initial_year = 2012, final_year = 2016)
}
\arguments{
\item{data}{The full Chicago crime dataset as given by \code{chicago}.}

\item{harmonics}{An integer in 1 to 6 specifiying the number of harmonics to model seasonality.}

\item{chains}{An integer specifying the number of MCMC chains for Stan to run.}

\item{iter}{An integer specifying the total number of samples for each chain, including warmup.}

\item{warmup}{An integer specifying the number of burn-in samples in Stan.}

\item{adapt_delta}{See Stan language manual at \url{http://mc-stan.org/users/documentation/}.}

\item{crime_types}{Character vector specifying the types of crimes to include in the analysis.
The function requires at least two crime types.}

\item{initial_year}{Integer in 2007 to 2015 specifying the first year in the analysis.}

\item{final_year}{Integer in 2008 to 2016 specifying the last year in the analysis.}
}
\value{
Returns a list of three objects: samples, summary, and crime_types. summary gives a summary of the
samples. crime_types returns the crime_types argument specified in the function. This serves as a reminder of
 the ordering of crime types, which is important for interpreting covariance matrix output 
 as well as making sure that the order of crime types is consistent between other functions.
 samples is itself a list where each element is either a two or three dimensional array
 containing all of the samples of a given parameter. The first dimension of the array always
 corresponds to the sample/iteration. Parameter names work as follows: 
 \code{omega_evo} corresponds to evolution correlation matrices,
 \code{omega_error} corresponds to error correlation matrices,
 \code{sigma_evo} corresponds to a vector of evolution standard deviations,
 \code{sigma_error} corresponds to a vector of the error standard deviations,
 \code{sigma_evo_mat} corresponds to evolution covariance matrices,
 \code{sigma_error_mat} corresponds to error covariance matrices.
 See Stan documentation for \code{lp__}.
}
\description{
\code{run_mcmc} uses Stan to run a Hamiltonian Monte Carlo algorithm to sample the observation
and latent state covariance parameters in the dynamic linear model specificied in the research narrative.
The full chicago crime dataset must be passed to the \code{data} argument. Other parameter arguments dictate
the subset of the data on which the analysis is performed as well as details of the MCMC to be passed to \code{rstan::sampling}.
}
\examples{
 \dontrun{
 cov_samples <- run_mcmc(data = chicago, chains = 2, adapt_delta = 0.8)
 }

}
